Simple clock

+++ src/io/ConsoleDevice.cpp
@@ -3,7 +3,7 @@
 #include "../efi.h"
 #include "../log.h"
 
-int ConsoleDevice::Write(uint16_t* message)
+int ConsoleDevice::Write(uchar* message)
 {
     g_EfiSystemTable->ConOut->OutputString(g_EfiSystemTable->ConOut, message);
     return 0;
