+++ src/io/ConsoleDevice.h
@@ -4,5 +4,5 @@ class ConsoleDevice : public IOutputDevice {
   public:
     ConsoleDevice() {}
 
-    virtual int Write(uint16_t* message);
+    virtual int Write(uchar* message);
 };
